[refactor] findAllColorOfMonth를 DiaryController -> ColorControler
