+++ src/main/java/com/emotionmaster/emolog/diary/service/DiaryService.java
@@ -28,8 +28,4 @@ public class DiaryService {
     public void delete(long id){
         diaryRepository.deleteById(id);
     }
-
-    public List<Diary> findAllColorOfMonth(int month){
-        return diaryRepository.findAllByMonth(month);
-    }
 }
